requirements.txt
Django>=5.0,<6
psycopg2-binary
discord.py>=2.3
pytz
asgiref
openpyxl
django-cors-headers
